requirements.txt
pytest
requests
pyjwt
pdoc
twine
wheel